Kcaptcha: Corrections for PHP 8.1 compatibility.

+++ platform/common/libraries/Kcaptcha.php
@@ -250,7 +250,7 @@ class Kcaptcha {
 
                                     for ($px = min($left, $width - 1); $px > $left - 200 && $px >= 0; $px -= 1) {
 
-                                        $color = imagecolorat($img, $px, $py) & 0xff;
+                                        $color = imagecolorat($img, (int) $px, (int) $py) & 0xff;
 
                                         if ($color + $opacity < 170) { // 170 - threshold
 
@@ -277,7 +277,7 @@ class Kcaptcha {
                     $shift = 1;
                 }
 
-                imagecopy($img, $font, $x-$shift, $y, $m['start'], 1, $m['end'] - $m['start'], $fontfile_height);
+                imagecopy($img, $font, (int) $x-$shift, (int) $y, (int) $m['start'], 1, (int) ($m['end'] - $m['start']), (int) $fontfile_height);
                 $x += $m['end'] - $m['start'] - $shift;
             }
 
@@ -339,10 +339,10 @@ class Kcaptcha {
 
                 } else {
 
-                    $color = imagecolorat($img, $sx, $sy) & 0xFF;
-                    $color_x = imagecolorat($img, $sx + 1, $sy) & 0xFF;
-                    $color_y = imagecolorat($img, $sx, $sy + 1) & 0xFF;
-                    $color_xy = imagecolorat($img, $sx + 1, $sy + 1) & 0xFF;
+                    $color = imagecolorat($img, (int) (int) $sx, (int) $sy) & 0xFF;
+                    $color_x = imagecolorat($img, (int) $sx + 1, (int) $sy) & 0xFF;
+                    $color_y = imagecolorat($img, (int) $sx, (int) $sy + 1) & 0xFF;
+                    $color_xy = imagecolorat($img, (int) $sx + 1, (int) $sy + 1) & 0xFF;
                 }
 
                 if ($color == 255 && $color_x == 255 && $color_y == 255 && $color_xy == 255) {
@@ -381,7 +381,7 @@ class Kcaptcha {
                     $newblue = $newcolor0*$foreground_color[2] + $newcolor*$background_color[2];
                 }
 
-                imagesetpixel($img2, $x, $y, imagecolorallocate($img2, $newred, $newgreen, $newblue));
+                imagesetpixel($img2, (int) $x, (int) $y, imagecolorallocate($img2, (int) $newred, (int) $newgreen, (int) $newblue));
             }
         }
